Refactor k8s aliases and add new functions
This commit refactors the k8s aliases in the aliases.sh file and adds new functions for better usability. The kdp alias is removed and replaced with a kdp function that allows for additional arguments. Additionally, new functions for docker compose are added: dcdu, dcd, dcu, and dcb. Lastly, new functions kat, katfwd, kat-open, and k8c are added for working with Kubernetes. The changes aim to improve the efficiency and convenience of working with k8s and docker compose.